fixed the bug in animation - maybe
